functions/gcf_process_splitter_results_source/requirements.txt
google-cloud-documentai==0.5.0
google-cloud-storage==1.37.1
PyPDF2==1.26.0
apyio==0.2.0
google-cloud-bigquery==2.20.0